Initial code commit - working
Has NewCaptureActivity, InProgressActivity and blank CaptureListActivity.

Records four lines of preamble prepended with the hash symbol ('#'), listing the room, a
detailed description of the position in the room, a description of the local activity and
the duration of the capture, respectively.

Each line of data has a timestamp in ISO-8601 format without the Z, assuming local timezone
in the Diamond on the date of capture, followed by the magnetic field strength in x, y and z
coordinates (in microTesla).